+++ src/Oro/Bundle/WebsiteSearchBundle/Engine/IndexerInputValidator.php
@@ -19,24 +19,13 @@ class IndexerInputValidator
 {
     use ContextTrait;
 
-    private WebsiteProviderInterface $websiteProvider;
-    private SearchMappingProvider $mappingProvider;
-    private ManagerRegistry $managerRegistry;
-    private ReindexationWebsiteProviderInterface $reindexationWebsiteProvider;
-    private TokenAccessorInterface $tokenAccessor;
-
     public function __construct(
-        WebsiteProviderInterface $websiteProvider,
-        SearchMappingProvider $mappingProvider,
-        ManagerRegistry $managerRegistry,
-        ReindexationWebsiteProviderInterface $provider,
-        TokenAccessorInterface $tokenAccessor
+        private WebsiteProviderInterface $websiteProvider,
+        private SearchMappingProvider $mappingProvider,
+        private ManagerRegistry $managerRegistry,
+        private ReindexationWebsiteProviderInterface $reindexationWebsiteProvider,
+        private TokenAccessorInterface $tokenAccessor
     ) {
-        $this->websiteProvider = $websiteProvider;
-        $this->mappingProvider = $mappingProvider;
-        $this->managerRegistry = $managerRegistry;
-        $this->reindexationWebsiteProvider = $provider;
-        $this->tokenAccessor = $tokenAccessor;
     }
 
     public function validateRequestParameters(array|string|null $classOrClasses, array $context): array
@@ -67,12 +56,14 @@ class IndexerInputValidator
             $resolver->setDefined(AbstractIndexer::CONTEXT_WEBSITE_IDS);
             $resolver->setDefined(AbstractIndexer::CONTEXT_ENTITIES_IDS_KEY);
             $resolver->setDefined(AbstractIndexer::CONTEXT_CURRENT_WEBSITE_ID_KEY);
+            $resolver->setDefined(AbstractIndexer::CONTEXT_BATCH_SIZE);
 
             $resolver->setAllowedTypes('skip_pre_processing', ['bool']);
             $resolver->setAllowedTypes(AbstractIndexer::CONTEXT_FIELD_GROUPS, ['string[]']);
             $resolver->setAllowedTypes(AbstractIndexer::CONTEXT_WEBSITE_IDS, ['int[]', 'string[]']);
             $resolver->setAllowedTypes(AbstractIndexer::CONTEXT_ENTITIES_IDS_KEY, ['int[]', 'string[]']);
             $resolver->setAllowedTypes(AbstractIndexer::CONTEXT_CURRENT_WEBSITE_ID_KEY, 'int');
+            $resolver->setAllowedTypes(AbstractIndexer::CONTEXT_BATCH_SIZE, 'int');
 
             $organization = $this->tokenAccessor->getOrganization();
             $resolver->setDefault(